Ensure relayer tests in ckb-sync package are free from known unstable factors
